+++ implementation/js/src/infrastructure/api.js
@@ -341,6 +341,30 @@
 			return window.cancelAnimationFrame(id);
 		},
 
+		//----------------markdown-------------------
+		markdown: function(md, $target, options){
+			if($target && ($target instanceof jQuery))
+				$target.html(marked(md, options));
+			else
+				return marked(md, options || $target);
+		},
+
+		//----------------notify---------------------
+		notify: function(title /*or options*/, msg, type /*or otherOptions*/, otherOptions){
+			if(_.isString(title))
+				$.amaran(_.extend({
+					content: {
+						themeName: 'stagejs',
+						title: title,
+						message: msg, 
+						type: (_.isString(type) && type) || 'info',
+					},
+					themeTemplate: app.NOTIFYTPL
+				}, otherOptions || type));
+			else
+				$.amaran(title);
+		},
+
 		//----------------debug----------------------
 		debug: function(){
 			var fn = console.debug || console.log;
@@ -368,7 +392,7 @@
 		'lock', 'unlock', 'available', //global action locks
 		'coop', 'navigate', 'reload', 'param', 'animation', 'nextFrame', 'cancelFrame',
 		'remote', 'ws', 'download', //com
-		'extract', 'cookie', 'store', 'moment', 'uri', 'validator', //3rd-party lib short-cut
+		'extract', 'cookie', 'store', 'moment', 'uri', 'validator', 'markdown', 'notify', //3rd-party lib short-cut
 		//@supportive
 		'debug', 'has', 'get', 'nameToPath', 'pathToName', 'inject.js', 'inject.tpl', 'inject.css',
 		//@deprecated
@@ -380,5 +404,7 @@
 	 */
 	//animation done events used in Animate.css
 	app.ADE = 'webkitAnimationEnd mozAnimationEnd MSAnimationEnd oanimationend animationend';
+	//notification template
+	app.NOTIFYTPL = Handlebars.compile('<div class="alert alert-dismissable alert-{{type}}"><button data-dismiss="alert" class="close" type="button">×</button><strong>{{title}}</strong> {{{message}}}</div>');
 
 })(Application);