+++ implementation/js/src/infrastructure/lib+-/marionette/layout.js
@@ -142,11 +142,22 @@
 			//find region marks after rendering and ensure region.$el (to support dynamic template)
 			this.listenTo(this, 'render', function(){
 				var that = this;
+				//a. named regions (for dynamic navigation)
 				$(this.el.outerHTML).find('[region]').each(function(index, el){
 					var r = $(el).attr('region');
 					//that.regions[r] = '[region="' + r + '"]';
 					that.regions[r] = {
 						selector: '[region="' + r + '"]'
+					};
+				});
+				//b. anonymous regions (for static view nesting)
+				$(this.el.outerHTML).find('[view]').each(function(index, el){
+					var $el = $(el);
+					if($el.attr('region')) return; //skip dynamic regions (already detected)
+
+					var r = $el.attr('view');
+					that.regions[_.uniqueId('anonymous-region-')] = {
+						selector: '[view="' + r + '"]'
 					};
 				});
 				this.addRegions(this.regions);